fixed potential uninitialized value of shotnoise seed
Shotnoise seed was uninitialized (most likely 0), when not explicitly
specified in &setup (in G4 infile) and not explicitly specified using
'-s' command line option. Now it should be 123456789 (default value
in the manual).

Before fixing there was a compiler warning:
warning: ‘seed’ may be used uninitialized in this function [-Wmaybe-uninitialized]